unit(test): header component
