+++ src/App.js
@@ -1,5 +1,5 @@
-import Footer from "components/shared/Footer";
-import Header from "components/shared/Header";
+import Footer from "components/shared/footer/Footer";
+import Header from "components/shared/header/Header";
 import Tracker from "components/tracker/Tracker";
 import "./App.scss";
